+++ libs/blocks/src/lib/navigation/language-switcher/index.tsx
@@ -1,6 +1,6 @@
 import { Text } from '@deriv/quill-design';
 import { useLanguageSwitcher, useNavigation } from '@deriv-com/hooks';
-import { LabelPairedGlobeRegularIcon } from '@deriv/quill-icons/LabelPaired';
+import { StandaloneGlobeRegularIcon } from '@deriv/quill-icons';
 
 const LanguageSwitcher = () => {
   const { onLanguageHover, activeLanguage } = useLanguageSwitcher();
@@ -15,7 +15,7 @@ const LanguageSwitcher = () => {
         onLanguageHover?.();
       }}
     >
-      <LabelPairedGlobeRegularIcon className="inline" iconSize="md" />
+      <StandaloneGlobeRegularIcon className="inline" />
       <Text
         size="sm"
         className="inline-block group-hover:text-typography-prominent"