+++ libs/blocks/src/lib/navigation/mobile/index.tsx
@@ -13,9 +13,12 @@ export const MobileNav: React.FC<NavigationProps> = ({
   topNavigation,
 }) => {
   return (
-    <FluidContainer className="fixed z-50 w-screen bg-background-primary-container lg:hidden">
+    <FluidContainer
+      className="fixed z-50 w-screen bg-background-primary-container lg:hidden"
+      data-cy="mobile-header"
+    >
       {topNavigation?.()}
-      <nav className="flex min-h-[80px] w-full flex-row items-center justify-between gap-gap-3xl">
+      <nav className="flex min-h-[80px] w-full flex-row items-center justify-between md:gap-gap-3xl">
         {renderLogo?.()}
         <NavMobileWrapper>
           <div className="flex h-full flex-col justify-between">