+++ libs/blocks/src/lib/navigation/mobile/mobile.nav-link.tsx
@@ -2,7 +2,7 @@ import { CustomLink } from '@deriv-com/components';
 import { useNavigation } from '@deriv-com/hooks';
 import { DirectNavLink } from '@deriv-com/providers';
 import { qtMerge } from '@deriv/quill-design';
-import { LabelPairedArrowUpRightRegularIcon } from '@deriv/quill-icons/LabelPaired';
+import { LabelPairedArrowUpRightSmRegularIcon } from '@deriv/quill-icons/LabelPaired';
 
 export interface MobileNavLinkProps {
   item: DirectNavLink;
@@ -20,6 +20,7 @@ export const MobileNavLink: React.FC<MobileNavLinkProps> = ({
       href={item.href}
       hasHoverDecoration={false}
       className={qtMerge(
+        'text-typography-default',
         'py-general-sm text-body-md font-bold leading-body-md',
         'hover:text-typography-prominent',
         className,
@@ -28,7 +29,7 @@ export const MobileNavLink: React.FC<MobileNavLinkProps> = ({
       target={item.target}
     >
       {item.text}
-      {item.external && <LabelPairedArrowUpRightRegularIcon iconSize="sm" />}
+      {item.external && <LabelPairedArrowUpRightSmRegularIcon />}
     </CustomLink>
   );
 };